Draft
We think it is a bit too long and some of the chapters get too much
into details, we are still missing some images (the foot fetish one and
a few more weird convos). And the headers or footnotes for the
comparisons. But overall as a first draft we are happy :). We also need
a strong conclusion and an intro which judging by the title i think it
should be more personal, any way it is good to have a 3rd eye maybe 4th
:).